test_gpu_clock.py: Move test_clock from `tests` to `benchmarks`
